requirements.txt
amqp==5.0.3
asgiref==3.3.1
beautifulsoup4==4.9.3
billiard==3.6.3.0
bootstrap4==0.1.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Deprecated==1.2.11
Django==3.1.5
django-bootstrap4==2.3.1
django-bower==5.2.0
django-currentuser==0.5.2
django-extensions==3.1.0
django-filter==2.4.0
django-js-asset==1.2.2
django-kanban-budget==0.1.4
django-model-utils==4.1.1
django-mptt==0.11.0
django-scheduler==0.9.3
django-scrumboard==0.1.1
django-upload-validator==1.1.5
django-uuslug==1.2.0
djangorestframework==3.12.2
icalendar==4.0.7
idna==2.10
kombu==5.0.2
libmagic==1.0
Markdown==3.3.3
Pillow==8.1.0
prompt-toolkit==3.0.14
pycparser==2.20
PyGithub==1.54.1
PyJWT==1.7.1
python-dateutil==2.8.1
#python-magic-bin==0.4.14
#python-magic-bin
python-magic
python-slugify==4.0.1
pytz==2020.5
requests==2.25.1
six==1.15.0
slacker==0.14.0
soupsieve==2.1
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
